requirements.txt
paho-mqtt
requests
python-socketio[client]